backend/pyproject.toml
[project]
name = "meeting-agent-backend"
version = "0.1.0"
description = "AI Meeting Agent - analyzes transcripts and generates calendar invites, decision records, and reports"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "openai",
    "faster-whisper",
    "icalendar",
    "google-api-python-client",
    "google-auth-oauthlib",
    "python-multipart",
    "python-dotenv",
    "sse-starlette",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
